reintroduce required images
