+++ csfieldguide/static/interactives/box-translation/js/box-translation.js
@@ -1,33 +1,9 @@
-var commonFunctions = require('../../../js/translation-rotation-interactives/common-functions.js');
+var commonFunctions = require('../../../js/translation-rotation-interactives/modular-functions.js');
 var TWEEN = require('@tweenjs/tween.js');
 
 scale = 10;
 
 module.exports = {
-
-    moveBox: function(x_pos, y_pos, z_pos, cube, render) {
-        /* triggered when the user hits the "enter" (or "return") key
-        * updates the cube's position to match the coordinates inputted by the user
-        */
-
-        commonFunctions.emptyCheck(x_pos, y_pos, z_pos);
-
-        // use the coordinates to set a new target for the box
-        var target = {
-            x: x_pos * scale,
-            y: y_pos * scale,
-            z: z_pos * scale
-        };
-
-        // move the box on the next animation loop
-        TWEEN.removeAll();
-        new TWEEN.Tween( cube.position )
-            .to( target, 1000 )
-            .easing( TWEEN.Easing.Back.InOut )
-            .onUpdate( render )
-            .start();
-
-    },
 
 
     /*
