+++ csfieldguide/static/js/translation-rotation-interactives/modular-functions.js
